[LCR-57] fix: rename conjestionDatas to ConjestionDatas
